Don't create a log directory every time the launch logger is imported. (#299)
* Don't create a log directory every time the launch logger is imported.

To do this, we revamp the launch_config() function into a class
with member variables and methods.  We then make the 'log_dir'
a property, and if it is None when it is accessed, we create
the directory then.  This ensures that we will not create the
directory until someone actually wants to use it.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* Switch to using a property setter for the level.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* Make log_dir a real property, and remove from configure.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* Make the log_handler_factory a property, removing from configure.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* Move screen_format/screen style to a getter/setter.

This gets it out of configure.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* Switch log_format to a setter method.

This allows us to remove configure.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* Review feedback.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>